requirements.txt
Flask
Flask-SQLAlchemy
boto3
python-dotenv
mysql-connector-python
